feat(pdf): enhance GOCA support with Bezier curves and fractional line widths
- Updated PdfGraphicsState to support float line widths for increased precision.
- Implemented GOCA drawing orders: GBSEG (recursive), GSFLW (fractional line width), GCBEZ/GCCBEZ (cubic Bezier curves).
- Added comprehensive unit and integration tests in PdfHandlerStructureTest and PdfGraphicsStateTest.
- Updated ADF_TO_PDF_ROADMAP.md to reflect current implementation status.

+++ src/main/java/com/mgz/pdf/PdfGraphicsState.java
@@ -31,7 +31,7 @@ import com.mgz.afp.goca.GAD_DrawingOrder.GSLJ_SetLineJoin.LineJoin;
 public class PdfGraphicsState {
 
   private AFPColorValue color = AFPColorValue.DeviceDefault_0x00;
-  private short lineWidth = 0;
+  private float lineWidth = 0.0f;
   private short lineType = 0;
   private LineEnd lineEnd = LineEnd.Default;
   private LineJoin lineJoin = LineJoin.Default;
@@ -65,7 +65,7 @@ public class PdfGraphicsState {
    */
   public void reset() {
     this.color = AFPColorValue.DeviceDefault_0x00;
-    this.lineWidth = 0;
+    this.lineWidth = 0.0f;
     this.lineType = 0;
     this.lineEnd = LineEnd.Default;
     this.lineJoin = LineJoin.Default;
@@ -101,11 +101,11 @@ public class PdfGraphicsState {
     this.color = color;
   }
 
-  public short getLineWidth() {
+  public float getLineWidth() {
     return lineWidth;
   }
 
-  public void setLineWidth(short lineWidth) {
+  public void setLineWidth(float lineWidth) {
     this.lineWidth = lineWidth;
   }
